requirements.txt
aiohttp==3.12.14
azure-identity==1.23.0
azure-appconfiguration==1.7.1
azure-monitor-opentelemetry==1.6.10
azure-monitor-opentelemetry-exporter==1.0.0b39
azure-storage-blob==12.25.1
chainlit==2.6.0
fastapi==0.115.14
httpx==0.28.1
itsdangerous==2.2.0
msal==1.32.3
opentelemetry-instrumentation-httpx==0.52b1
python-dotenv==1.1.1
starlette==0.46.2
tenacity==9.1.2
uvicorn==0.35.0
azure-appconfiguration-provider==2.1.0
azure-applicationinsights==0.1.1